refactor: rename util to utils to match backend
